Fix large amount of new tests
