+++ src/tests/functional/plugin/conformance/test_runner/conformance_infra/src/skip_tests_config.cpp
@@ -22,6 +22,8 @@ std::vector<std::string> disabledTests = {};
 
 ov::AnyMap pluginConfig = {};
 
+ShapeMode shapeMode = ov::test::conformance::ShapeMode::BOTH;
+
 } // namespace conformance
 } // namespace test
 } // namespace ov
