--- src/tests/functional/plugin/conformance/test_runner/op_conformance_runner/include/read_ir_test/read_ir_tests.hpp
@@ -1,21 +1,0 @@
-// Copyright (C) 2018-2023 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-//
-
-#include "read_ir_test/read_ir.hpp"
-
-namespace ov {
-namespace test {
-namespace subgraph {
-
-TEST_P(ReadIRTest, ReadIR) {
-    run();
-}
-
-TEST_P(ReadIRTest, QueryModel) {
-    query_model();
-}
-
-} // namespace subgraph
-} // namespace test
-} // namespace ov
