+++ src/tests/functional/plugin/conformance/test_runner/op_conformance_runner/src/op_impl_check/op_impl_check_tests.cpp
@@ -1,0 +1,30 @@
+// Copyright (C) 2018-2023 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#include "op_impl_check/op_impl_check_compile_model.hpp"
+#include "op_impl_check/op_impl_check_query_model.hpp"
+#include "op_impl_check/single_op_graph.hpp"
+#include "conformance.hpp"
+
+namespace ov {
+namespace test {
+namespace conformance {
+namespace op {
+
+using namespace ov::test::subgraph;
+
+namespace {
+INSTANTIATE_TEST_SUITE_P(conformance,
+                         OpImplCheckTest,
+                         ::testing::Combine(
+                                 ::testing::ValuesIn(createFunctions()),
+                                 ::testing::Values(targetDevice),
+                                 ::testing::Values(pluginConfig)),
+                         OpImplCheckTest::getTestCaseName);
+} // namespace
+
+}  // namespace op
+}  // namespace conformance
+}  // namespace test
+}  // namespace ov
