+++ src/tests/functional/plugin/conformance/test_runner/op_conformance_runner/src/op_impl_check/single_op_graph.cpp
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include <single_layer_tests/op_impl_check/op_impl_check.hpp>
-#include <single_layer_tests/op_impl_check/single_op_graph.hpp>
+#include <op_impl_check/op_impl_check.hpp>
+#include <op_impl_check/single_op_graph.hpp>
 #include <openvino/pass/constant_folding.hpp>
 
 namespace ov {
