+++ src/tests/ie_test_utils/functional_test_utils/layer_tests_summary/template/report_template.html
@@ -32,6 +32,9 @@
                 <span><b>Passrates are based on relative weights each subgraphs! You can check absolute value in `General passrate` row!</b></span>
             </div>
             <div>
+                <span><b>Relative Passrate could be from 0 to 200% now! Relative weight of operation is calculated once for current set of models, but there are more then one tests per model. </b></span>
+            </div>
+            <div>
                 <span><b>Status:</b></span>
                 <span class="green">P:85</span><span>Passed</span>
                 <span class="red">F:0</span><span>Failed</span>
